extract LineAccumulator class
